requirements/production.txt
certifi==2017.4.17
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.2
django-filter==1.0.4
djangorestframework==3.6.3
drf-tracking==1.2.0
factory-boy==2.8.1
Faker==0.7.18
idna==2.5
Markdown==2.6.8
prettyconf==1.2.3
psycopg2==2.7.1
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.2
six==1.10.0
urllib3==1.22
whitenoise==3.3.0
httpie
pycrypto==2.6.1  
django-mptt==0.8.7
drfdocs==0.0.11
django-markdownx
coreapi
pygments
